Avoid re-downloading existing GIS files

+++ src/p2a_analyses_read_processed_data.R
@@ -108,9 +108,11 @@ data <- data[-which(data$name == "Ben" & data$dt == as.POSIXct(strptime("2016-11
 ################################################
 library(rgdal);library(sp);library(raster)
 
-download.file("https://www.naturalearthdata.com/http//www.naturalearthdata.com/download/50m/cultural/ne_50m_admin_0_countries.zip","data/maps/ne_50m_admin_0_countries.zip")
-unzip(zipfile = "data/maps/ne_50m_admin_0_countries.zip", exdir = "data/maps/ne_50m_admin_0_countries")
-file.remove("data/maps/ne_50m_admin_0_countries.zip")
+if(!dir.exists("data/maps/ne_50m_admin_0_countries")) {
+  download.file("https://www.naturalearthdata.com/http//www.naturalearthdata.com/download/50m/cultural/ne_50m_admin_0_countries.zip","data/maps/ne_50m_admin_0_countries.zip")
+  unzip(zipfile = "data/maps/ne_50m_admin_0_countries.zip", exdir = "data/maps/ne_50m_admin_0_countries")
+  file.remove("data/maps/ne_50m_admin_0_countries.zip")
+}
 
 countries <- shapefile(here("data", "maps", "ne_50m_admin_0_countries/ne_50m_admin_0_countries.shp"))
 
@@ -134,5 +136,3 @@ colnames(out)[3] <- "continent"
 data <- merge(data,out[,c("name","dt","continent")],all.x=T)
 data$continent <- ifelse(is.na(data$continent) == T,'sea',as.character(data$continent))
 data$continent <- factor(as.factor(data$continent),levels=c("Europe","Africa","sea"))
-
-
